requirements.txt
alembic==1.7.7
bcrypt==3.1.7
Flask==2.1.0
Flask-Bcrypt==0.7.1
Flask-SQLAlchemy==2.5.1
flask_uploads==0.2.1
Flask-WTF==0.14.3
itsdangerous==1.1.0
click==8.1.0
cffi==1.15.0
Flask-Login==0.5.0
Flask-Migrate==3.1.0
pdfkit==0.6.1
pycparser==2.19
six==1.14.0
vine==1.3.0
Werkzeug~=2.0.0
wkhtmltopdf==0.2
WTForms==2.2.1
flask_msearch==0.2.5
Flask-SQLAlchemy==1.3.13
Flask-WTF==1.0.0
Faker==13.3.4
pymysql==1.0.2